requirements.txt
aiogram==3.3.0
aiohttp
requests
beautifulsoup4
python-dotenv
fastapi
uvicorn
natasha
razdel
pymorphy2
pymorphy2-dicts-ru
lxml